Migrate diagnostic pipeline
